remove nesting for Billing component folder and move it to be directly under src/components instead
