+++ src/pages/billings/view/index.js
@@ -1,4 +1,4 @@
-import ViewBillingDetailCard from 'components/Student/Billing/ViewBillingDetailCard'
+import ViewBillingDetailCard from 'components/Billing/ViewBillingDetailCard'
 import React from 'react'
 import { useParams } from 'react-router-dom'
